node: Add documentation and cleanup for tests

+++ node-appliance/tests/testImage.py
@@ -1,5 +1,24 @@
 #!/usr/bin/env python
-# vim: et ts=4 sw=4 sts=4
+# -*- coding: utf-8 -*-
+#
+# Copyright 2015 Red Hat, Inc.
+#
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 2 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
+#
+# Refer to the README and COPYING files for full details of the license
+#
 
 import unittest
 import sh
@@ -8,7 +27,6 @@ import tempfile
 import os
 import shutil
 import re
-from unittest import SkipTest
 
 
 # Increase the capture length of python-sh to show complete errors
@@ -17,12 +35,6 @@ sh.ErrorReturnCode.truncate_cap = 999999
 
 def log(x):
     print(x)
-
-
-def lines(data):
-    """This method is not suited for big data
-    """
-    return [l for l in data.split("\n") if l]
 
 
 squashfsimg = os.environ.get("TEST_NODE_SQUASHFS_IMG",
@@ -37,10 +49,12 @@ class TestRootfsQcow2Image(unittest.TestCase):
     sh = _fish.bake("sh")
 
     def test_package(self):
+        """Ensure the main packages are installed
+        """
         req_pkgs = ["vdsm", "cockpit"]
 
         missing_pkgs = []
-        all_builds = lines(self.sh("rpm -qa"))
+        all_builds = self.sh("rpm -qa").splitlines()
 
         def name_from_build(build):
             return re.sub(r"-[^-]+-[^-]+$", "", build)
@@ -56,12 +70,16 @@ class TestRootfsQcow2Image(unittest.TestCase):
             "Some packages are missing: %s" % missing_pkgs
 
     def test_partitioning(self):
-        pt = lines(self._fish("list-partitions"))
+        """Ensure that there is only one partition in the appliance image
+        """
+        pt = self._fish("list-partitions").splitlines()
         pt_len = len(pt)
         assert pt_len == 1, "More than one partition: %s" % pt_len
 
     def test_selinux_denials(self):
-        denials = lines(self.sh("grep denied /var/log/audit/audit.log"))
+        """Ensure that there are no denials righ after appliance creation
+        """
+        denials = self.sh("grep denied /var/log/audit/audit.log").splitlines()
         assert len(denials) == 0, \
             "To many denials: %s\n%s" % (len(denials), denials)
 
@@ -84,19 +102,22 @@ class TestRootfsSquashfsImage(unittest.TestCase):
 
     @classmethod
     def tearDownClass(cls):
-        log("Removing tree: %s" % cls.dest)
-        assert os.getcwd() in cls.dest
-        shutil.rmtree(cls.dest)
+        if os.path.exists(cls.dest):
+            log("Removing tree: %s" % cls.dest)
+            assert os.getcwd() in cls.dest
+            shutil.rmtree(cls.dest)
 
     def test_one_partition(self):
+        """Ensure that the image is really just a filesystem image
+
+        … and not a disk image (image with a partition table).
+        """
         from sh import file as _file
         filetype = _file(self.img)
         log("Found filetype: %s" % filetype)
         assert "filesystem" in filetype
 
     def test_fsck(self):
+        """Ensure that the filesystem is clean
+        """
         sh.fsck("-nvf", self.img)
-
-    def test_valid_fstree(self):
-        raise SkipTest("Not implemented")
-        # FIXME check if /boot /sbin etc exists
